+++ packages/article-byline/src/article-byline-prop-types.js
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import { Text } from "react-native";
-import { treePropType } from "@times-components/markup";
+import { propTypes as treePropType } from "@times-components/markup-forest";
 
 export const propTypes = {
   ast: PropTypes.arrayOf(treePropType).isRequired,