+++ packages/article/__tests__/comments.native.js
@@ -73,6 +73,7 @@ export default () => {
         onLinkPress={() => {}}
         onRelatedArticlePress={() => {}}
         onTopicPress={() => {}}
+        onTwitterLinkPress={() => {}}
         onVideoPress={() => {}}
       />
     );