+++ packages/pull-quote/__tests__/shared.base.js
@@ -1,6 +1,8 @@
 import React from "react";
+import { shallow } from "enzyme";
 import { iterator } from "@times-components/test-utils";
 import PullQuotes from "../src/pull-quote";
+import PullQuoteTwitterLink from "../src/pull-quote-twitter-link";
 
 const content = "Some content";
 const caption = "A caption";
@@ -12,7 +14,11 @@ export default renderComponent => {
       name: "with a caption",
       test() {
         const output = renderComponent(
-          <PullQuotes caption={caption} content={content} />
+          <PullQuotes
+            caption={caption}
+            content={content}
+            onTwitterLinkPress={() => null}
+          />
         );
 
         expect(output).toMatchSnapshot();
@@ -21,7 +27,9 @@ export default renderComponent => {
     {
       name: "without a caption",
       test() {
-        const output = renderComponent(<PullQuotes content={content} />);
+        const output = renderComponent(
+          <PullQuotes content={content} onTwitterLinkPress={() => null} />
+        );
 
         expect(output).toMatchSnapshot();
       }
@@ -30,10 +38,32 @@ export default renderComponent => {
       name: "with a twitter handle",
       test() {
         const output = renderComponent(
-          <PullQuotes caption={caption} content={content} twitter={twitter} />
+          <PullQuotes
+            caption={caption}
+            content={content}
+            onTwitterLinkPress={() => null}
+            twitter={twitter}
+          />
         );
 
         expect(output).toMatchSnapshot();
+      }
+    },
+    {
+      name: "twitter handle can be clicked",
+      test() {
+        const onTwitterLinkPressMock = jest.fn();
+
+        const wrapper = shallow(
+          <PullQuoteTwitterLink
+            onTwitterLinkPress={onTwitterLinkPressMock}
+            twitter={twitter}
+          />
+        );
+
+        wrapper.find("TextLink").simulate("press", {});
+
+        expect(onTwitterLinkPressMock).toHaveBeenCalled();
       }
     }
   ];