+++ android-app/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.7.7"></a>
+## [0.7.7](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.7.6...@times-components/android-app@0.7.7) (2018-08-17)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+<a name="0.7.6"></a>
+## [0.7.6](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.7.5...@times-components/android-app@0.7.6) (2018-08-16)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+<a name="0.7.5"></a>
+## [0.7.5](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.7.4...@times-components/android-app@0.7.5) (2018-08-16)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
 <a name="0.7.4"></a>
 ## [0.7.4](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.7.3...@times-components/android-app@0.7.4) (2018-08-16)