+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.1.1"></a>
+## [1.1.1](https://github.com/newsuk/times-components/compare/@times-components/provider@1.1.0...@times-components/provider@1.1.1) (2018-08-16)
+
+**Note:** Version bump only for package @times-components/provider
+
+
+
+
+
 <a name="1.1.0"></a>
 # [1.1.0](https://github.com/newsuk/times-components/compare/@times-components/provider@1.0.5...@times-components/provider@1.1.0) (2018-08-15)